Refactor to move utils to llm_utils
